Removed the bug that the AJAX loading animation won't go away.
Added a failure handler for AJAX requests.

Refactored the way debug messages are added. Instead of Opentip.debug() you now better use this.debug() since it adds the Tip ID to the debug message in a nice standard way.

Added debug information to the AJAX calls.

OK! That's it... next time I'll split it up in more commits. I swear!